# Conflicts:
#	src/main/java/com/shophelperjavasb/shophelperjavasb/products/model/Product.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/products/repositories/ProductsRepository.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/users/controllers/UsersController.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/users/controllers/api/AuthApi.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/users/controllers/api/UsersApi.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/users/model/User.java
#	src/main/java/com/shophelperjavasb/shophelperjavasb/users/services/UsersService.java

+++ src/main/java/com/shophelperjavasb/shophelperjavasb/purchases/repositories/PurchasesRepository.java
@@ -3,5 +3,5 @@ package com.shophelperjavasb.shophelperjavasb.purchases.repositories;
 import com.shophelperjavasb.shophelperjavasb.purchases.model.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PurchasesRepository extends JpaRepository<Purchase, Integer> {
+public interface PurchasesRepository extends JpaRepository<Purchase, Long> {
 }
